Added parent child resolvers to Variant allele
